+++ src/core/requests/headers.py
@@ -212,6 +212,8 @@ def check_http_traffic(request):
     http_response_content(page)
     # Checks regarding a potential CAPTCHA protection mechanism.
     checks.captcha_check(page)
+    # Checks regarding a potential browser verification protection mechanism.
+    checks.browser_verification(page)
 
   # This is useful when handling exotic HTTP errors (i.e requests for authentication).
   except urllib2.HTTPError, err: